Update AdventofCodeDay2.py
I originally wrote this code with my classes containing get and set functions, which is how I was trained classically mostly in Java. I was told it is both standard and Pythonic to not include functions and rather reference variables directly. So for this change, I wanted to update this code to look less like Java and more like Python